[GAME-5] Game Designer Control Options
Set options of combinations and density of recipes that are generated. Options to set available as slider in manager's inspector.

Item density: chance 0 - 100 for each grid to spawn an item
Recipe 2 ing range: Sets the amount of 2 ingredients recipes on play
Recipe 3 ing range: Sets the amount of 3 ingredients recipes on play
Recipe 4 ing range: Sets the amount of 4 ingredients recipes on play

Configuration can be saved or load in a file.